requirements.txt
ase==3.22.1
bqplot==0.12.40
bqplot_image_gl==1.4.11
cupy==12.2.0
easygraphics==2.0.1.2
gpaw==23.6.1
h5py==3.7.0
hyperspy==1.7.5
imageio==2.27.0
ipywidgets==7.6.3
matplotlib==3.5.3
mpi4py==3.1.4
numba==0.56.4
numpy==1.18.5
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
Pillow==9.5.0
Pillow==10.0.0
psutil==5.9.0
psutil==5.9.4
pyFFTW==0.12.0
pytest==7.3.0
pytorch_lightning==1.5.6
PyYAML==6.0
PyYAML==6.0.1
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.4.1
scipy==1.7.3
setuptools==65.6.3
setuptools==67.6.1
sympy==1.12
tensorflow_gpu==2.3.0
torch==1.13.0+cu116
torchsummary==1.5.1
torchvision==0.14.0+cu116
tqdm==4.65.0
traitlets==5.9.0
traittypes==0.2.1
